Create README.md
This repository hold and contains the IOS Code for our FSE301 Project. 

Please do not ever commit to master, unless a pull request is submitted to allow for the code to be properly merged. 

You can make a separate branch to allow for your own commits of code. Please make sure when doing so that you are committing to your branch and not master.